+++ zbus/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] – 2023-09-30
+
 ### Added
 - Re-export `logcontrol` and `logcontrol::DBUS_OBJ_PATH`.
 - Add `logcontrol_zbus::ConnectionBuilderExt` to extend `ConnectionBuilder` with `serve_log_control`.
@@ -19,5 +21,6 @@ Initial release.
 
 - Add DBus interface implementation `LogControl1`.
 
-[Unreleased]: https://github.com/swsnr/logcontrol.rs/compare/logcontrol-zbus-v1.0.0...HEAD
+[Unreleased]: https://github.com/swsnr/logcontrol.rs/compare/logcontrol-zbus-v1.1.0...HEAD
+[1.1.0]: https://github.com/swsnr/logcontrol.rs/compare/logcontrol-zbus-v1.0.0...logcontrol-zbus-v1.1.0
 [1.0.0]: https://github.com/swsnr/logcontrol.rs/releases/tag/logcontrol-zbus-v1.0.0